Rename loading component file
